backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
redis==5.2.1
celery==5.4.0
httpx==0.28.1
stripe==11.4.1
slowapi==0.1.9
structlog==24.4.0
websockets==14.1
coqui-tts==0.25.3
ffmpeg-python==0.2.0
pillow==11.0.0
